+++ src/main/resources/db/migration/V001__initialize_schema_and_tables.sql
@@ -9,6 +9,7 @@ CREATE TABLE samply.project
     state_machine_key TEXT,
     name              TEXT,
     state             TEXT,
+    creator_email     TEXT,
     expiration_date   DATE,
     created_at        TIMESTAMP,
     expires_at        TIMESTAMP,